blog/requirements.txt
aliyun-python-sdk-core==2.13.15
Django==2.2
django-ranged-response==0.2.0
django-redis==4.11.0
django-simple-captcha==0.5.12
django-tinymce==2.8.0
image==1.5.28
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
vc==2018.7.10
verifycode==0.2.1
